Made Wearable non-abstract, changed name to Equipment. Updated all references to it. Added type variable to Equipment
