--- src/mjearlb/game/items/wearables/Helmet.java
@@ -1,9 +1,0 @@
-package mjearlb.game.items.wearables;
-
-public class Helmet extends Wearable {
-
-    public Helmet(String name, int defense) {
-	super(name, defense); 
-    } // Helmet
-
-} // Helmet
